Port part of KCalUtils to Grantlee templates
Instead of assembling the HTML by hand in C++ code, we use Grantlee templates
for presentation, and only prepare the data in C++ code. This commit only ports
the event, tasks, journal and freebusy views to Grantlee templates. The ITIP
HTML code is still assembled in C++, I want to wait with that until Sandro is
finished with his work on ObjectTreeParser refactoring in kdepim.git to see what
the requierements will be.

This code also includes a bunch of useful Grantlee plugins and tools that we
 might want to eventually move to some other repository so it can be used
from other projects as well.

REVIEW: 125331