Remove KDateTime::Spec from the IncidenceFormatter API
This is always used with the local time zone anyway, as it's for
displaying content to the user.

There is one deviation in the test data caused by this, but that actually
seems correct, DTSTART;TZID=Europe/Prague:20150930T030000 is 01:00:00 UTC.

This also fixes both the incidence view and the calendar view tooltips
showing content of varying correctness for recurring events in different
timezones.