Fix autotests/testincidenceformatter
At the moment testincidenceformatter passes in the PIM Docker image but
fails in Jenkins because of differences in palette colors for button
backgrounds, foregrounds, and borders.

Testing for specific colors has repeatedly proved to be fragile.  This
patch abstracts them out, and tests that the colors used are whatever the
current palette provides.