IncidenceFormatter: fix start/end dt of recurrent event crossing DST
This fixes a particular case when a recurrent event is created during
DST but a subsequent recurrence occurs after DST ends. The formatter
would still show the start/end time of the occurrence as if the
DST still applied. This is best noticed when system timezone does not
observe DST, but the displayed incidence from a timezone uses DST.

This change replaces a custom (broken) algorithm by the one provided
by KCalendarCore to obtain start/end datetime of an incidence on
a particular date, which is actually correct.

BUG: 451459
BUG: 460186
FIXED-IN: 5.24.1